+++ test/standalone/Vegetation/plant_hydraulics_test.jl
@@ -115,7 +115,8 @@ for FT in (Float32, Float64)
 
         AR_params = AutotrophicRespirationParameters(FT)
         RTparams = BeerLambertParameters(FT)
-        photosynthesis_params = FarquharParameters(FT, C3())
+        is_c3 = FT(1) # set the photosynthesis mechanism to C3
+        photosynthesis_params = FarquharParameters(FT, is_c3)
         stomatal_g_params = MedlynConductanceParameters(FT)
 
         AR_model = AutotrophicRespirationModel{FT}(AR_params)
@@ -403,7 +404,8 @@ for FT in (Float32, Float64)
 
         AR_params = AutotrophicRespirationParameters(FT)
         RTparams = BeerLambertParameters(FT)
-        photosynthesis_params = FarquharParameters(FT, C3())
+        is_c3 = FT(1) # set the photosynthesis mechanism to C3
+        photosynthesis_params = FarquharParameters(FT, is_c3)
         stomatal_g_params = MedlynConductanceParameters(FT)
 
         AR_model = AutotrophicRespirationModel{FT}(AR_params)